refactoring, added slime and mob-related abstractions

+++ Assets/Scripts/Mobs/DropInfo.cs
@@ -3,5 +3,5 @@ public struct DropInfo
 {
     public int ItemId;
     public int ItemCount;
-    public int DropChance;
+    public float DropChance;
 }
